Refs. #7. Add TextCat plugin.
Implement the a plugin that tries to detect
the language of the message. The plugin has
several options to tweak which languages are
accepted and how probabile they have to be
to be taken into consideration.

Expose an eval rule that performes the actual
check and exposes the data in the message.

Some of the options of the SA plugin are not
needed in our implementation, so they are just
put there as placeholders.