Change functional daemon tests to set the user config in setup
The daemon will read the user configuration once and then store the resulting
ruleset in memory to save resources. Adjust the user override test to set the
user config in the setUpClass method. And check that the score is actually
changed when the user option is sent to the daemon.

Also add a separate test that checks that after the daemon is reloaded, with
the SIGUSR1 signal, the new configuration is read from the user preferences
and the rule score is overridden.

Reference #44